take the api outside app.vue
